+++ containers/catalog/catalog.page.jsx
@@ -1,0 +1,14 @@
+import { getLayout as mainPageLayout } from 'components/layout/main/main.component'
+import PageLayout from 'components/layout/main/page/page.component'
+
+const Catalog = () => {
+  return (
+    <PageLayout title='Catálogo'>
+
+    </PageLayout>
+  )
+}
+
+Catalog.getLayout = mainPageLayout
+
+export default Catalog
