+++ containers/monetization/monetization.page.jsx
@@ -1,0 +1,14 @@
+import { getLayout as mainPageLayout } from 'components/layout/main/main.component'
+import PageLayout from 'components/layout/main/page/page.component'
+
+const Monetization = () => {
+  return (
+    <PageLayout title='Pedidos'>
+
+    </PageLayout>
+  )
+}
+
+Monetization.getLayout = mainPageLayout
+
+export default Monetization
