+++ containers/users/users.page.jsx
@@ -1,0 +1,14 @@
+import { getLayout as mainPageLayout } from 'components/layout/main/main.component'
+import PageLayout from 'components/layout/main/page/page.component'
+
+const Users = () => {
+  return (
+    <PageLayout title='Usuários'>
+
+    </PageLayout>
+  )
+}
+
+Users.getLayout = mainPageLayout
+
+export default Users
