+++ pages/catalog/index.jsx
@@ -1,14 +1,3 @@
-import { getLayout as mainPageLayout } from 'components/layout/main/main.component'
-import PageLayout from 'components/layout/main/page/page.component'
+import CatalogPage from 'containers/catalog/catalog.page'
 
-const Catalog = () => {
-  return (
-    <PageLayout title='Catálogo'>
-
-    </PageLayout>
-  )
-}
-
-Catalog.getLayout = mainPageLayout
-
-export default Catalog
+export default CatalogPage
