+++ pages/monetization/index.jsx
@@ -1,14 +1,3 @@
-import { getLayout as mainPageLayout } from 'components/layout/main/main.component'
-import PageLayout from 'components/layout/main/page/page.component'
-
-const Monetization = () => {
-  return (
-    <PageLayout title='Pedidos'>
-
-    </PageLayout>
-  )
-}
-
-Monetization.getLayout = mainPageLayout
+import Monetization from 'containers/monetization/monetization.page'
 
 export default Monetization
