+++ pages/users/index.jsx
@@ -1,14 +1,3 @@
-import { getLayout as mainPageLayout } from 'components/layout/main/main.component'
-import PageLayout from 'components/layout/main/page/page.component'
+import UsersPage from 'containers/users/users.page'
 
-const Users = () => {
-  return (
-    <PageLayout title='Usuários'>
-
-    </PageLayout>
-  )
-}
-
-Users.getLayout = mainPageLayout
-
-export default Users
+export default UsersPage
